Finally
I got the basics I wanted in this program to work, in a brief moment of
inspiration I remembered the properties of .append and the for loops
(python version of it). Next set of features I am hoping to bring to
the table: add in human elements of error, and possibly multi-stack
capabilities (meaning I would have to simplify the procedure and use
layers of procedures). Also something cool would be the saving of a
list so that it can be further mutated.